+++ ava.config.js
@@ -1,0 +1,12 @@
+
+"use strict";
+
+export default {
+    "compileEnhancements": false,
+    "extensions": [
+        "ts"
+    ],
+    "require": [
+        "ts-node/register"
+    ]
+};